Fix for post load event regression
If page.js is loaded after the `load` event fires, the `onpopstate`
function is never attached to the `popstate` event.

This tweak prevents that, but also prevents the hack which fixes #213
from working if page.js is loaded after the load event fires.

I think this is an acceptable compromise for now, because it allows
#213 to remain fixed if page.js is loaded before the `load` event fires.

We can always replace this with something better if someone identifies
a solution which fixes it in all cases.