Change test infrastructure to run inside of iframes
This moves individual test to all run inside of iframes. This ensures
that there is some level of isolation for tests and that you can run
them in an actual browser and not just depend on jsdom.

This fixes #442

It also provides a new way to use an alternative to the main window by
passing the `window` option when starting page.

```js
page('/about', function(){

});

page({
  window: myIframe.contentWindow
});
```

In the future (probably post-2.0) I want to make it possible to have
multiple Pages running at the same time, but for now this is what we'll
have.