Ensure we are on the same page when exiting for hashes
The onclick listening tries not to do anything when navigating to a hash
within the same page. However previously it did not take the `search`
into consideration so when you were on a page with a query param and
then navigating to the same path, without a query param, but with a
hash, it would cause a page navigation.

Fix is to include the search in this algorithm. Fixes #289