Add page.create() to allow multiple pages
This adds a new feature, `page.create()` that allows for there to be
multiple pages. Calling `page.create(options)` create a new page
function that behaves just like the global page. It contains its own
internal state, options, and event handlers. You can target the page to
another window, like an iframe or a popup, if you want.

```js
var newPage = page.create({
      window: someIframe.contentWindow
      });
```

The main reason for adding this feature was to clean up or testing
infrastructure. Numerous things had to be done in the past to make sure
each test was cleanly in isolation. That stuff is no longer necessary
now that we can just create a new page instance for each test.

Closes #13