Implement dispatching iteratively.
In applications with a large number of middlewares, the recursive
implementation of nextEnter and nextExit become a problem, by
generating deep call stacks (2x the number of middlewares, as far
as I can tell).

Even where the call stack size isn't directly an issue, the added
cruft to the call tree can make debugging and profiling harder.

There doesn't seem to be a strong reason for these methods to be
implemented recursively, so I reimplemented them iteratively.