add routerId to config options: support multiple instances of page
When an application has multiple heroku apps mounted under the same domain and each app uses PageJS as its client-side router, the browser back/forward buttons don't work between the seperate apps. This is because the popstate event handler of each instance assumes that it is the only entity modifing the history state, and therefore tries to route paths that it may not own.

For this reason, I added the option for a routerId to disinguish between multiple apps (that each use a different instance of page). Each instance now includes its id in the history state, and that id is used by page to know whether to route on the client's side, or defer to the server.