Prevent duplicate event handler registration by configuring the router only once
Remove the `options` parameter from `page.start`/`createPage` (it wasn't used anyway) and
also from the `Page` constructor together with the `configure` call there.

That ensures that the router is configured only once, with the user-supplied `options`,
instead of being configured twice: first with default `options` in constructor, second time
with user-supplied `options` in `page.start`.

Fixes a real-world issue where passing `{ click: false }` to `page.start` didn't prevent
adding the global click handler to the document anyway, because the default value of the
`click` option is `true`.